Server 2:nd version
New functions for handling new users and sending messages.

New disconnect function that removes them from the userlist.

New message-types:
s:userListUpdate: the server sends a copy of the userlist to each of the clients so that the list on the page is always up-to-date.

s:newUserData: Replaces the s:allMessages-type since the server sends these when a user connects. It includes both the chat-history as well as the newly generated username.

Lots of comments to (hopefully) make the code more understandable.