Fixed correct usernames
Made it so that the assigned username is saved in a global variable and that that name gets put as the sender when sending a message to the server.
When the client recieves a message it checks if the senders name is the same as the name saved in the global variable. If it is the sender gets put as "You" and the class "ownmMessage" is added. Otherwise the sender gets put as the username that was sent with the message.